[Console] Add support for error ouput in the CommandTester
