+++ src/Symfony/Component/Console/CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
    pass it the command as an array of its arguments instead
  * made the `ProcessHelper` class final
  * added `WrappableOutputFormatterInterface::formatAndWrap()` (implemented in `OutputFormatter`)
+ * added `capture_stderr_separately` option to `CommandTester::execute()`
 
 4.1.0
 -----
